Sub List links rendering VIA extension
The sublist link renders now via the extension.
In the extension define a sub_list view or any other named view
and in the helper function of the extenstion pass it through.
This allows for extensions to handle their own nav. Weather that be
a list or a tree type list.