Added the new BackButton to model views/lists
Also refactored the create method for lists a bit.  There is a bug in
Enyo where kind properties that are arrays (or objects) are shared by
all instances of that kind that the refactor works around.  It also
switches the create method to the more performant enyo.inherit
overriding syntax.